image_edit/requirements.txt
git+https://github.com/huggingface/diffusers.git



transformers
accelerate
safetensors
sentencepiece
dashscope
kernels
torchvision
peft
torchao==0.11.0
av